+++ CHANGELOG.md
@@ -1,16 +1,22 @@
 # Changelog
 
+## v0.4.0 (2021-05-06)
+
+- Add: Wildcard and RegExp match for `Ignore Files`
+- Add: `.*` (Dot files) as default for `Ignore Files`
+- Optimize: Set note title without file extension #6
+
 ## v0.3.3 (2021-03-16)
 
 - Fix: typo
 
 ## v0.3.2 (2021-02-22)
 
-Fix: File is not deleted on tagging error
+- Fix: File is not deleted on tagging error
 
 ## v0.3.1 (2021-01-29)
 
-Fix: Folder with a beginning dot is excluded #5
+- Fix: Folder with a beginning dot is excluded #5
 
 ## v0.3.0 (2021-01-29)